chore: move i18n files around
