+++ src/components/inputs/async-creatable-select-input/async-creatable-select-input.js
@@ -11,7 +11,7 @@ import addStaticFields from '../../../utils/add-static-fields';
 import LoadingIndicator from '../../internals/loading-indicator';
 import ClearIndicator from '../../internals/clear-indicator';
 import DropdownIndicator from '../../internals/dropdown-indicator';
-import messages from './messages';
+import messages from '../../internals/messages/select';
 import createSelectStyles from '../../internals/create-select-styles';
 
 const useTheme = () => useContext(ThemeContext);
