--- src/components/inputs/select-input/messages.js
@@ -1,19 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  noOptionsMessageWithInputValue: {
-    id: 'UIKit.SelectInput.noOptionsMessageWithInputValue',
-    description: 'Text of dropdown when no options match search text',
-    defaultMessage: 'No options',
-  },
-  noOptionsMessageWithoutInputValue: {
-    id: 'UIKit.SelectInput.noOptionsMessageWithoutInputValue',
-    description: 'Text of dropdown when no options exist',
-    defaultMessage: 'No options',
-  },
-  placeholder: {
-    id: 'UIKit.SelectInput.placeholder',
-    description: 'Default input placeholder text',
-    defaultMessage: 'Select...',
-  },
-});
