+++ src/components/internals/messages/select/index.js
@@ -1,0 +1,1 @@
+export { default } from './messages';
